Extract type checker utils file
